css/css-anchor-position/anchor-scope-dynamic.html

<!DOCTYPE html>
<title>CSS Anchor Positioning: anchor-name is contained by anchor-scope:all - dynamic containment</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#determining">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
  .container {
    position: relative;
  }
  .anchor {
    width: 100px;
    height: 100px;
    anchor-name: var(--anchor-name);
    background-color: skyblue;
  }
  .anchor-scope-all {
    anchor-scope: all;
  }
  .target {
    position: absolute;
    position-anchor: var(--anchor-name);
    top: anchor(bottom, 50px);
  }
  #a1 { --anchor-name: --a1; }
  #a2 { --anchor-name: --a2; }
  #a3 { --anchor-name: --a3; }
</style>
<div id="a1" class="container">
  <div class="anchor"></div>
  <div class="anchor-scope-all">
    <div class="anchor"></div>
  </div>
  <div class="target" data-offset-y="100"></div>
</div>
<div id="a2" class="container">
  <div class="anchor"></div>
  <div class="anchor"></div>
  <div class="anchor-scope-all">
    <div class="target" data-offset-y="50"></div>
  </div>
</div>
<div id="a3" class="container">
  <div class="anchor anchor-scope-all">
    <div class="target" data-offset-y="50"></div>
  </div>
</div>
<script type="module">
  await checkLayoutForAnchorPos('.target');

  const t1 = document.querySelector("#a1 .target");
  const t2 = document.querySelector("#a2 .target");
  const t3 = document.querySelector("#a3 .target");
  const t4 = document.querySelector("#a4 .target");
  const t5 = document.querySelector("#a5 .target");
  t1.setAttribute("data-offset-y", "200");
  t2.setAttribute("data-offset-y", "200");
  t3.setAttribute("data-offset-y", "100");
  for (let element of document.querySelectorAll(".anchor-scope-all")) {
    element.style.anchorScope = "initial";
  }
  await checkLayoutForAnchorPos('.target');

  t1.setAttribute("data-offset-y", "100");
  t2.setAttribute("data-offset-y", "50");
  t3.setAttribute("data-offset-y", "50");
  for (let element of document.querySelectorAll(".anchor-scope-all")) {
    element.style.anchorScope = "";
  }
  await checkLayoutForAnchorPos('.target');
</script>
